fix check not checking completed series for already released novels, added interactive name selection for series to remove, and handled float values for hanashi order json entries

+++ magnum/internal/hanashimedia/json.go
@@ -9,7 +9,7 @@ type JSONVolumeInfo struct {
 		Description string `json:"description"`
 		Completed   bool   `json:"completed"`
 		Entries     []struct {
-			Order int `json:"order"`
+			Order float32 `json:"order"`
 			Ebook struct {
 				At           string    `json:"at"`
 				Title        string    `json:"title"`
